feat: add cardinality check on arbitrary set generation (#130)
This pull request updates the arbitrary types to include a new property called "supremumCardinality" which calculates the maximum number of possible values for the arbitrary. This change affects various arbitrary types such as tuple, string, set, object, array, boolean, char, constants, date, and integer. The new property is used to provide better estimations for the maximum number of generated values in a given context.

---